refactor to extract layout and code gen logic
